% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ecm.R
\name{add_logo}
\alias{add_logo}
\title{Adicionar logo do Economia Mainstream
Esta função adiciona o logo do Economia Mainstream no canto inferior esquerdo}
\usage{
add_logo(logo, position_x = 0.07, position_y = -0.03, width_set = 1.5)
}
\arguments{
\item{position_x}{Localização do logo em relação ao eixo X. Pré-definido para 0.07}

\item{position_y}{Localização do logo em relação ao eixo Y. Pré-definido para -0.03}

\item{width_set}{Define a largura do logo em inches. Pré-definido em 1.5}
}
\value{
Adiciona o logo do ECM no canto inferior esquerdo do Plot.
}
\description{
Adicionar logo do Economia Mainstream
Esta função adiciona o logo do Economia Mainstream no canto inferior esquerdo
}
