% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ecm.R
\name{theme_ecm}
\alias{theme_ecm}
\title{Tema do Economia Mainstream
Esta função gera o tema do Economia Mainstream. Questão de fonte etc..
Voce não deveria se preocupar com fontes e etc.}
\usage{
theme_ecm(
  base_size = 12,
  base_family = "Montserrat",
  laranja = "#fa7c01",
  cinza = "#2f2f2f",
  branco = "#ffffff",
  preto = "#1d1d1d"
)
}
\arguments{
\item{base_size}{Tamanho de fonte}
}
\value{
Tema do ECM
}
\description{
Tema do Economia Mainstream
Esta função gera o tema do Economia Mainstream. Questão de fonte etc..
Voce não deveria se preocupar com fontes e etc.
}
